Move findUnexpectedExtraInput to ParserUtil
