+++ src/main/java/seedu/address/logic/parser/Parser.java
@@ -1,8 +1,5 @@
 package seedu.address.logic.parser;
 
-import java.util.Locale;
-import java.util.Optional;
-
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.parser.exceptions.ParseException;
 
@@ -16,37 +13,4 @@ public interface Parser<T extends Command> {
      * @throws ParseException if {@code userInput} does not conform the expected format
      */
     T parse(String userInput) throws ParseException;
-
-    /**
-     * Returns the first disallowed prefixed token in input order, if any.
-     * A prefix is only recognized when preceded by whitespace, matching ArgumentTokenizer behavior.
-     */
-    static Optional<String> findUnexpectedExtraInput(String args, Prefix[] disallowedPrefixes) {
-        int earliestPosition = -1;
-        String unexpectedToken = null;
-
-        for (Prefix prefix : disallowedPrefixes) {
-            int position = findPrefixPosition(args, prefix);
-            if (position != -1 && (earliestPosition == -1 || position < earliestPosition)) {
-                earliestPosition = position;
-                unexpectedToken = extractToken(args, position);
-            }
-        }
-
-        return Optional.ofNullable(unexpectedToken);
-    }
-
-    private static int findPrefixPosition(String args, Prefix prefix) {
-        int prefixIndex = args.toLowerCase(Locale.ROOT)
-                .indexOf(" " + prefix.getPrefix().toLowerCase(Locale.ROOT));
-        return prefixIndex == -1 ? -1 : prefixIndex + 1;
-    }
-
-    private static String extractToken(String args, int startPosition) {
-        int endPosition = args.indexOf(' ', startPosition);
-        if (endPosition == -1) {
-            return args.substring(startPosition);
-        }
-        return args.substring(startPosition, endPosition);
-    }
 }
